refact of team generation structure
