+++ Docker/Dockerfile
@@ -7,6 +7,7 @@ RUN mkdir -p /app
 WORKDIR /app
 
 COPY ./package.json ./package.json
+COPY ./yarn.lock ./yarn.lock
 
 RUN yarn add -dev typescript@${TYPESCRIPT_VERSION}
 RUN yarn install
