+++ src/teamGeneration/v.0.2/types/Rank.ts
@@ -1,0 +1,4 @@
+export type Rank = {
+    name: string,
+    seed: number
+}
