+++ src/teamGeneration/v0.1/teamGeneration.ts
@@ -1,5 +1,5 @@
-import { Participant } from '../models/participant';
-import { Rank } from '../models/rank';
+import { Participant } from '../../models/participant';
+import { Rank } from '../../models/rank';
 
 // TEAM GENRERATION V0.1.r by Danaen
 /* eslint-disable */
